Move BooksListContainer to container folder
